chore(ui): re-organize visuals for empty state component
