+++ ui/src/assets/empty/visuals/apps.png
