+++ ui/src/assets/empty/visuals/concurrency_executions.png
