+++ ui/src/assets/empty/visuals/concurrency_limit.png
