+++ ui/src/assets/empty/visuals/dependencies.png
