+++ ui/src/assets/empty/visuals/triggers.png
